+++ core/shared/src/main/scala/cats/effect/IOAppCommon.scala
@@ -16,4 +16,8 @@
 
 package cats.effect
 
-trait IOAppCommon {}
+import cats.effect.unsafe.IORuntime
+
+trait IOAppCommon {
+  private[effect] def defaultGlobalRuntime: IORuntime
+}
